+++ modules/polling/helpers/utils.ts
@@ -1,6 +1,3 @@
-import { getArbitrumGoerliTestnetSdk, getArbitrumOneSdk } from '@dethcrypto/eth-sdk-client';
-import { config } from 'lib/config';
-import { ArbitrumSdkGenerators } from 'modules/web3/types/contracts';
 import { PollInputFormat, PollResultDisplay, PollVictoryConditions } from '../polling.constants';
 import { Poll, PollParameters, PollVictoryConditionAND, PollTallyVote, VictoryCondition } from '../types';
 
@@ -122,13 +119,3 @@ export function extractCurrentPollVote(
 export function findPollById(pollList: Poll[], pollId: string): Poll | undefined {
   return pollList.find((poll: Poll) => parseInt(pollId) === poll.pollId);
 }
-
-export const relayerCredentials = {
-  mainnet: { apiKey: config.DEFENDER_API_KEY_MAINNET, apiSecret: config.DEFENDER_API_SECRET_MAINNET },
-  goerli: { apiKey: config.DEFENDER_API_KEY_TESTNET, apiSecret: config.DEFENDER_API_SECRET_TESTNET }
-};
-
-export const arbitrumSdkGenerators: ArbitrumSdkGenerators = {
-  mainnet: getArbitrumOneSdk,
-  goerli: getArbitrumGoerliTestnetSdk
-};
